add ability to recover courses, sections, lessons

+++ app/controllers/lessons_controller.rb
@@ -39,8 +39,16 @@ class LessonsController < ApplicationController
 
   def update
     @lesson = Lesson.with_deleted.find(params[:id])
-    if params[:deleted]
-      restore_lesson(@lesson)
+    if params[:deleted] && params[:section_id]
+      @lesson.restore
+      lesson_section = LessonSection.only_deleted.find_by(lesson: @lesson, section_id: params[:section_id])
+      lesson_section.restore
+      redirect_to course_section_path(lesson_section.section.course, lesson_section.section), notice: "Lesson restored to this section."
+    elsif params[:deleted]
+      @lesson.restore
+      lesson_sections = LessonSection.only_deleted.where(lesson_id: @lesson.id)
+      lesson_sections.each { |lesson_section| lesson_section.restore }
+      redirect_to lesson_path(@lesson), notice: 'Lesson restored to all its sections.'
     else
       lesson_section_ids = lesson_params[:lesson_sections_attributes].map { |param| param[1][:id] }
       lesson_sections_to_delete = LessonSection.where(lesson: @lesson) - LessonSection.find(lesson_section_ids)
@@ -55,10 +63,14 @@ class LessonsController < ApplicationController
 
   def destroy
     lesson = Lesson.find(params[:id])
-    lesson_sections = LessonSection.where(lesson_id: lesson.id)
-    lesson.destroy
-    lesson_sections.each { |lesson_section| lesson_section.update(deleted_at: Time.zone.now) }
-    redirect_to courses_path, notice: 'Lesson deleted.'
+    if params[:section_id]
+      section = Section.find(params[:section_id])
+      LessonSection.find_by(lesson: lesson, section: section).destroy
+      redirect_to course_section_path(section.course, section), notice: 'Lesson removed from this section.'
+    else
+      lesson.destroy
+      redirect_to courses_path, notice: 'Lesson deleted.'
+    end
   end
 
 private
@@ -68,13 +80,6 @@ private
     params.require(:lesson).permit(:name, :content, :cheat_sheet, :update_warning, :teacher_notes,
                                    :public, :deleted_at, :video_id,
                                    lesson_sections_attributes: [:id, :work_type, :section_id, :day_of_week])
-  end
-
-  def restore_lesson(lesson)
-    lesson.restore!
-    lesson_sections = LessonSection.where(lesson_id: lesson.id)
-    lesson_sections.each { |lesson_section| lesson_section.update(deleted_at: nil) }
-    redirect_to lesson_path(lesson), notice: 'Lesson restored.'
   end
 
   def find_search_results
@@ -87,7 +92,7 @@ private
 
   def find_deleted_lessons
     @lessons = Lesson.only_deleted
-    lesson_sections = LessonSection.where(lesson_id: @lessons.map(&:id))
+    lesson_sections = LessonSection.only_deleted.where(lesson_id: @lessons.map(&:id))
     sections = Section.where(id: lesson_sections.map(&:section_id))
     @courses = sections.map { |section| Course.find(section.course_id) }.uniq
     render :deleted
